Ripped out FlattenException into a separate extension method
