+++ src/Nancy/Helpers/ExceptionExtensions.cs
@@ -1,0 +1,32 @@
+﻿namespace Nancy.Helpers
+{
+    using System;
+
+    internal static class ExceptionExtensions
+    {
+        internal static Exception FlattenInnerExceptions(this Exception exception)
+        {
+            var aggregateException = exception as AggregateException;
+            if (aggregateException != null)
+            {
+                var flattenedAggregateException = aggregateException.Flatten();
+
+                //If we have more than one exception in the AggregateException
+                //we have to send all exceptions back in order not to swallow any exceptions.
+                if (flattenedAggregateException.InnerExceptions.Count > 1)
+                {
+                    return flattenedAggregateException;
+                }
+
+                return flattenedAggregateException.InnerException.FlattenInnerExceptions();
+            }
+
+            if (exception != null && exception.InnerException != null)
+            {
+                return exception.InnerException.FlattenInnerExceptions();
+            }
+
+            return exception;
+        }
+    }
+}
